Clean up clutter in single .c file.
Split up into the current supported update architectures, now calling
v0 / v1 instead of ts7970/common.

Use compatible strings for detecting the different boards.
