+++ crc8.h
@@ -1,0 +1,3 @@
+#pragma once
+
+uint8_t crc8(uint8_t *input_str, size_t num_bytes);
