+++ meson.build
@@ -1,2 +1,11 @@
 project('tssupervisorupdate', 'c')
-executable('tssupervisorupdate', 'tssupervisorupdate.c', install : true)
+executable('tssupervisorupdate', 
+  [
+    'tssupervisorupdate.c',
+    'micro.c',
+    'update-v0.c',
+    'update-v1.c',
+	'crc8.c',
+  ], 
+  install : true
+)
